LPD-85380 Simplify, every license integration test should check if it is running on a release bundle
